+++ handshake-state.js
@@ -4,62 +4,62 @@ const assert = require('nanoassert')
 const clone = require('clone')
 const cipherState = require('./cipher-state')
 
-module.exports = ({dh, symmetricState}) => {
+module.exports = ({ dh, symmetricState }) => {
   var DhResult = sodium_malloc(dh.DHLEN)
 
   function HandshakeState () {
     this.symmetricState = sodium_malloc(symmetricState.STATELEN)
-  
+
     this.initiator = null
-  
+
     this.spk = null
     this.ssk = null
-  
+
     this.epk = null
     this.esk = null
-  
+
     this.rs = null
     this.re = null
-  
+
     this.messagePatterns = null
   }
 
-  function initialize(handshakePattern, initiator, prologue, s, e, rs, re) {
+  function initialize (handshakePattern, initiator, prologue, s, e, rs, re) {
     assert(Object.keys(PATTERNS).includes(handshakePattern), 'Unsupported handshake pattern')
     assert(typeof initiator === 'boolean', 'Initiator must be a boolean')
     assert(prologue.byteLength != null, 'prolouge must be a Buffer')
-  
+
     assert(e == null ? true : e.publicKey.byteLength === dh.PKLEN, `e.publicKey must be ${dh.PKLEN} bytes`)
     assert(e == null ? true : e.secretKey.byteLength === dh.SKLEN, `e.secretKey must be ${dh.SKLEN} bytes`)
-  
+
     assert(rs == null ? true : rs.byteLength === dh.PKLEN, `rs must be ${dh.PKLEN} bytes`)
     assert(re == null ? true : re.byteLength === dh.PKLEN, `re must be ${dh.PKLEN} bytes`)
-  
+
     var state = new HandshakeState()
-  
-    var protocolName = Uint8Array.from(`Noise_${handshakePattern}_25519_ChaChaPoly_BLAKE2b`, toCharCode)
-  
+
+    var protocolName = Uint8Array.from(`Noise_${handshakePattern}_${dh.ALG}_ChaChaPoly_BLAKE2b`, toCharCode)
+
     symmetricState.initializeSymmetric(state.symmetricState, protocolName)
     symmetricState.mixHash(state.symmetricState, prologue)
-  
+
     state.role = initiator === true ? INITIATOR : RESPONDER
-  
+
     if (s != null) {
       assert(s.publicKey.byteLength === dh.PKLEN, `s.publicKey must be ${dh.PKLEN} bytes`)
       assert(s.secretKey.byteLength === dh.SKLEN, `s.secretKey must be ${dh.SKLEN} bytes`)
-  
+
       state.spk = sodiumBufferCopy(s.publicKey)
       state.ssk = sodiumBufferCopy(s.secretKey)
     }
-  
+
     if (e != null) {
       assert(e.publicKey.byteLength === dh.PKLEN)
       assert(e.secretKey.byteLength === dh.SKLEN)
-  
+
       state.epk = sodiumBufferCopy(e.publicKey)
       state.esk = sodiumBufferCopy(e.secretKey)
     }
-  
+
     if (rs != null) {
       assert(rs.byteLength === dh.PKLEN)
       state.rs = sodiumBufferCopy(rs)
@@ -68,13 +68,13 @@ module.exports = ({dh, symmetricState}) => {
       assert(re.byteLength === dh.PKLEN)
       state.re = sodiumBufferCopy(re)
     }
-  
+
     // hashing
     var pat = PATTERNS[handshakePattern]
-  
+
     for (var pattern of clone(pat.premessages)) {
       var patternRole = pattern.shift()
-  
+
       for (var token of pattern) {
         switch (token) {
           case TOK_E:
@@ -90,18 +90,18 @@ module.exports = ({dh, symmetricState}) => {
         }
       }
     }
-  
+
     state.messagePatterns = clone(pat.messagePatterns)
-  
+
     assert(state.messagePatterns.filter(p => p[0] === INITIATOR).some(p => p.includes(TOK_S))
       ? (state.spk !== null && state.ssk !== null)
       : true, // Default if none is found
     'This handshake pattern requires a static keypair')
-  
+
     return state
   }
 
-  function writeMessage(state, payload, messageBuffer) {
+  function writeMessage (state, payload, messageBuffer) {
     assert(state instanceof HandshakeState)
     assert(payload.byteLength != null)
     assert(messageBuffer.byteLength != null)
@@ -189,49 +189,49 @@ module.exports = ({dh, symmetricState}) => {
     assert(state instanceof HandshakeState)
     assert(message.byteLength != null)
     assert(payloadBuffer.byteLength != null)
-  
+
     var mpat = state.messagePatterns.shift()
     var moffset = 0
-  
+
     assert(mpat != null)
     assert(mpat.shift() !== state.role)
-  
+
     for (var token of mpat) {
       switch (token) {
         case TOK_E:
           assert(state.re == null)
           assert(message.byteLength - moffset >= dh.PKLEN)
-  
+
           // PKLEN instead of DHLEN since they are different in out case
           state.re = sodium_malloc(dh.PKLEN)
           state.re.set(message.subarray(moffset, moffset + dh.PKLEN))
           moffset += dh.PKLEN
-  
+
           symmetricState.mixHash(state.symmetricState, state.re)
-  
+
           break
-  
+
         case TOK_S:
           assert(state.rs == null)
           state.rs = sodium_malloc(dh.PKLEN)
-  
+
           var bytes = 0
           if (symmetricState._hasKey(state.symmetricState)) {
             bytes = dh.PKLEN + 16
           } else {
             bytes = dh.PKLEN
           }
-  
+
           assert(message.byteLength - moffset >= bytes)
-  
+
           symmetricState.decryptAndHash(
             state.symmetricState,
             state.rs,
             message.subarray(moffset, moffset + bytes) // <- called temp in noise spec
           )
-  
+
           moffset += symmetricState.decryptAndHash.bytesRead
-  
+
           break
         case TOK_EE:
           dh.dh(DhResult, state.esk, state.re)
@@ -241,61 +241,60 @@ module.exports = ({dh, symmetricState}) => {
         case TOK_ES:
           if (state.role === INITIATOR) dh.dh(DhResult, state.esk, state.rs)
           else dh.dh(DhResult, state.ssk, state.re)
-  
+
           symmetricState.mixKey(state.symmetricState, DhResult)
           sodium_memzero(DhResult)
           break
         case TOK_SE:
           if (state.role === INITIATOR) dh.dh(DhResult, state.ssk, state.re)
           else dh.dh(DhResult, state.esk, state.rs)
-  
+
           symmetricState.mixKey(state.symmetricState, DhResult)
           sodium_memzero(DhResult)
           break
         case TOK_SS:
           dh.dh(DhResult, state.ssk, state.rs)
-  
+
           symmetricState.mixKey(state.symmetricState, DhResult)
           sodium_memzero(DhResult)
           break
-  
+
         default:
           throw new Error('Invalid message pattern')
       }
     }
-  
+
     symmetricState.decryptAndHash(state.symmetricState, payloadBuffer, message.subarray(moffset))
-  
+
     // How many bytes were written to payload (minus the TAG/MAC)
     readMessage.bytes = symmetricState.decryptAndHash.bytesWritten
-  
+
     if (state.messagePatterns.length === 0) {
       var tx = sodium_malloc(cipherState.STATELEN)
       var rx = sodium_malloc(cipherState.STATELEN)
       symmetricState.split(state.symmetricState, rx, tx, dh.DHLEN, dh.PKLEN)
-  
+
       return { tx, rx }
     }
   }
   readMessage.bytes = 0
-  
+
   function keygen (obj, sk) {
     if (!obj) {
       obj = { publicKey: sodium_malloc(dh.PKLEN), secretKey: sodium_malloc(dh.SKLEN) }
       return keygen(obj)
     }
-  
+
     if (obj.publicKey) {
       dh.generateKeypair(obj.publicKey, obj.secretKey)
       return obj
     }
-  
+
     if (obj.byteLength != null) dh.generateKeypair(null, obj)
   }
 
-
   function seedKeygen (seed) {
-    var obj = { publicKey: sodium_malloc(PKLEN), secretKey: sodium_malloc(SKLEN) }
+    var obj = { publicKey: sodium_malloc(dh.PKLEN), secretKey: sodium_malloc(dh.SKLEN) }
     dh.generateSeedKeypair(obj.publicKey, obj.secretKey, seed)
     return obj
   }
@@ -308,7 +307,7 @@ module.exports = ({dh, symmetricState}) => {
     keygen,
     seedKeygen,
     SKLEN: dh.SKLEN,
-    PKLEN: dh.PKLEN,
+    PKLEN: dh.PKLEN
   })
 }
